refactor: call 도메인 패키지 리팩터링
