--- domain/domain-pos/src/main/java/domain/pos/call/entity/vo/CallCategory.java
@@ -1,12 +1,0 @@
-package domain.pos.call.entity.vo;
-
-public enum CallCategory {
-	WATER,
-	PLATE,
-	TISSUE,
-	WATER_TISSUE,
-	SPOON,
-	CHOPSTICKS,
-	CUP,
-	OTHER
-}
